Vertex type changed from Long to Integer (64-bit to 32-bit).

+++ src/algorithm/AbstractMDSAlgorithm.java
@@ -5,7 +5,7 @@ import java.util.LinkedHashSet;
 import model.Graph;
 
 public interface AbstractMDSAlgorithm {
-	public LinkedHashSet<Long> mdsAlg(Graph g);
+	public LinkedHashSet<Integer> mdsAlg(Graph g);
 	public long getLastPrepTime();
 	public long getLastRunTime();
 }
