+++ src/algorithm/AbstractMSCAlgorithm.java
@@ -6,6 +6,6 @@ import java.util.LinkedHashSet;
 import model.Graph;
 
 public interface AbstractMSCAlgorithm {
-	public LinkedHashSet<Long> getMSCforMDS(LinkedHashSet<Long> universum,
+	public LinkedHashSet<Integer> getMSCforMDS(LinkedHashSet<Integer> universum,
 			ArrayList<RepresentedSet> sets, Graph g);
 }
