+++ src/algorithm/LessByN1AComparator.java
@@ -4,11 +4,11 @@ import java.util.Comparator;
 
 import model.Graph;
 
-public class LessByN1AComparator implements Comparator<Long> {
+public class LessByN1AComparator implements Comparator<Integer> {
 	private Graph g;
 
 	@Override
-	public int compare(Long arg0, Long arg1) {
+	public int compare(Integer arg0, Integer arg1) {
 		int is = g.getN1(arg0).size();
 		int js = g.getN1(arg1).size();
 		return is - js;
