+++ src/algorithm/chapter7/Algorithm34Vertex.java
@@ -9,23 +9,23 @@ import model.Graph;
 
 public class Algorithm34Vertex extends Thread {
 	private Graph g;
-	private LinkedHashSet<Long> W;
-	private Set<Long> S;
-	private Set<Long> G;
-	private Long v;
+	private LinkedHashSet<Integer> W;
+	private Set<Integer> S;
+	private Set<Integer> G;
+	private Integer v;
 	private Object lock;
 	private Object waitingForSpans = new Object();
-	private Map<Long, Long> spans;
-	private LinkedHashMap<Long, Algorithm34Vertex> instances;
+	private Map<Integer, Integer> spans;
+	private LinkedHashMap<Integer, Algorithm34Vertex> instances;
 	private final Object canJoinLock = new Object();
-	private Long canJoin = 0L; // mutex - how many unfinished joins are in
+	private Integer canJoin = 0; // mutex - how many unfinished joins are in
 								// neighbourhood
-	private LinkedHashSet<Long> dist2notSorG;
+	private LinkedHashSet<Integer> dist2notSorG;
 	private final Object wLock = new Object();
-	private Long w;
+	private Integer w;
 
-	public Algorithm34Vertex(Graph g, Long v, Set<Long> S, Set<Long> G,
-			LinkedHashMap<Long, Algorithm34Vertex> instances, Object lock) {
+	public Algorithm34Vertex(Graph g, Integer v, Set<Integer> S, Set<Integer> G,
+			LinkedHashMap<Integer, Algorithm34Vertex> instances, Object lock) {
 		this.g = g;
 		this.v = v;
 		this.S = S;
@@ -38,8 +38,8 @@ public class Algorithm34Vertex extends Thread {
 		this.w = computeSpan(v);
 	}
 
-	private boolean hasWhiteNeighbours(Long v) {
-		Set<Long> A;
+	private boolean hasWhiteNeighbours(Integer v) {
+		Set<Integer> A;
 		synchronized (W) {
 			A = new LinkedHashSet<>(W);
 		}
@@ -47,15 +47,15 @@ public class Algorithm34Vertex extends Thread {
 		return !A.isEmpty();
 	}
 
-	private Long computeSpan(Long v) {
-		Long result;
+	private Integer computeSpan(Integer v) {
+		Integer result;
 		synchronized (W) {
-			result = Long.valueOf(W.size());
+			result = Integer.valueOf(W.size());
 		}
 		return result;
 	}
 
-	public void recieveSpan(Long from, Long span) {
+	public void recieveSpan(Integer from, Integer span) {
 		synchronized (spans) {
 			spans.put(from, span);
 			// System.out.println("span recieved from " + from + " at " + v +
@@ -66,7 +66,7 @@ public class Algorithm34Vertex extends Thread {
 		}
 	}
 
-	public void recieveRemoveFromW(Long from) {
+	public void recieveRemoveFromW(Integer from) {
 		synchronized (W) {
 			W.remove(from);
 			// System.out.println("W remove recieved from " + from + " at " +
@@ -77,27 +77,27 @@ public class Algorithm34Vertex extends Thread {
 		}
 	}
 
-	public void recieveJoinS(Long from) {
+	public void recieveJoinS(Integer from) {
 		synchronized (dist2notSorG) {
 			dist2notSorG.remove(from);
 		}
 	}
 
-	public void recieveJoinG(Long from) {
+	public void recieveJoinG(Integer from) {
 		synchronized (dist2notSorG) {
 			dist2notSorG.remove(from);
 		}
 	}
 
-	private void joinS(Long v, Graph g) {
+	private void joinS(Integer v, Graph g) {
 		synchronized (dist2notSorG) {
 			System.out.print(v + " <-- ");
 			System.out.print(dist2notSorG);
-			for (Long v2 : dist2notSorG) {
+			for (Integer v2 : dist2notSorG) {
 				instances.get(v2).takeABreakPlease();
 				// System.out.println(instances.get(v2).getW());
 			}
-			for (Long v2 : dist2notSorG) {
+			for (Integer v2 : dist2notSorG) {
 				Algorithm34Vertex alg = instances.get(v2);
 				alg.recieveRemoveFromW(v);
 				// alg.recieveSpan(v, 0L);
@@ -110,7 +110,7 @@ public class Algorithm34Vertex extends Thread {
 				S.add(v);
 			}
 			System.out.println(" --> " + v);
-			for (Long v2 : dist2notSorG) {
+			for (Integer v2 : dist2notSorG) {
 				instances.get(v2).thankYou();
 			}
 			dist2notSorG.remove(v);
@@ -118,11 +118,11 @@ public class Algorithm34Vertex extends Thread {
 		return;
 	}
 
-	private void joinG(Long v) {
+	private void joinG(Integer v) {
 		synchronized (G) {
 			G.add(v);
 		}
-		for (Long v2 : dist2notSorG) {
+		for (Integer v2 : dist2notSorG) {
 			Algorithm34Vertex alg = instances.get(v2);
 			alg.recieveRemoveFromW(v);
 			// alg.recieveSpan(v, 0L);
@@ -147,7 +147,7 @@ public class Algorithm34Vertex extends Thread {
 		}
 	}
 
-	public Set<Long> getW() {
+	public Set<Integer> getW() {
 		return W;
 	}
 
@@ -166,7 +166,7 @@ public class Algorithm34Vertex extends Thread {
 				w = computeSpan(v); // span
 
 				synchronized (dist2notSorG) {
-					for (Long v2 : dist2notSorG) {
+					for (Integer v2 : dist2notSorG) {
 						instances.get(v2).recieveSpan(v, w);
 					}
 				}
@@ -181,14 +181,14 @@ public class Algorithm34Vertex extends Thread {
 				}
 
 				// if (spans.keySet().containsAll(dist2notSorG)) {
-				LinkedHashMap<Long, Long> oldspans = new LinkedHashMap<>();
+				LinkedHashMap<Integer, Integer> oldspans = new LinkedHashMap<>();
 				oldspans.putAll(spans);
 				w = computeSpan(v);
 
 				boolean isBiggest = true;
 				synchronized (dist2notSorG) {
-					for (Long v2 : dist2notSorG) {
-						Long getV2 = spans.get(v2);
+					for (Integer v2 : dist2notSorG) {
+						Integer getV2 = spans.get(v2);
 						if ((getV2 > w)
 								|| ((getV2.equals(w)) && (v2 < v))) {
 							isBiggest = false;
